Extract Person model to models folder
